refactor: Rename NovelCategory component
